Clean up for MELPA
Saw the comment in the README:

  I’d love to put this in Melpa, but I haven’t figured out how to do
  that.

As one of the first steps when you submit something to MELPA, the
checklist will be to run `package-lint` against your package to see if
there are any issues. If there are to address the first.

So this is what I've done here, which are:

- adding of the correct dependancy declaration, here emacs 24 as a
  requirement for lexical-bindings

- I've renamed all function names with org-slack to ox-slack, as this
  is one of the requirement of the package format for MELPA

- I've cleaned up some formatting, like when there were trailing
  parentheses or leading spaces before comments

I think that the rest should be addressed by the package author based
on his tastes.